requirements.txt
Pillow
huggingface_hub
